+++ app/routes/base64.tsx
@@ -1,14 +1,20 @@
 import * as b64 from "base64-encoding";
-import EncoderDecoder from "~/components/encoder-decoder";
-import { useCallback, useState } from "react";
+import { EncoderDecoderOutput } from "~/components/encoder-decoder-output";
+import { useCallback, useMemo, useState } from "react";
 import { metaHelper } from "~/utils/meta";
 import { utilities } from "~/utilities";
 import { BoxOptions } from "~/components/box";
+import { Utiliti } from "~/components/utiliti";
 
 export const meta = metaHelper(
   utilities.base64.name,
   utilities.base64.description
 );
+
+enum Action {
+  Encode = "Encode",
+  Decode = "Decode",
+}
 
 async function decode(text: string): Promise<string> {
   try {
@@ -22,6 +28,15 @@ async function decode(text: string): Promise<string> {
 
 export default function Base64() {
   const [urlSafe, setUrlSafe] = useState(false);
+
+  const renderOutput = useCallback(
+    (a: string, input: string, output: string) => {
+      if (a === Action.Encode || a === Action.Decode) {
+        return <EncoderDecoderOutput output={output} />;
+      }
+    },
+    []
+  );
 
   const encode = useCallback(
     async (text: string) => {
@@ -38,13 +53,31 @@ export default function Base64() {
     [urlSafe]
   );
 
+  const actions = useMemo(
+    () => ({
+      Encode: (input: string) => encode(input),
+      Decode: (input: string) => decode(input),
+    }),
+    [decode, encode]
+  );
+
   return (
-    <EncoderDecoder
+    <Utiliti
       label="Base64"
-      encode={encode}
-      decode={decode}
+      actions={actions}
+      renderInput={(input, setInput) => (
+        <textarea
+          id="input"
+          rows={3}
+          className="block font-mono w-full px-3 py-2 text-sm border-0 bg-zinc-800 focus:ring-0 text-white placeholder-zinc-400"
+          placeholder="Paste in your content…"
+          required={true}
+          value={input}
+          onChange={(e) => setInput(e.target.value)}
+        ></textarea>
+      )}
+      renderOutput={renderOutput}
       showLoadFile={true}
-      rows={10}
       renderOptions={() => (
         <BoxOptions>
           <div className="flex items-center h-5 w-5 ml-2">
